#3/feat/config pkgs restructured
