+++ internal/config/dashboard.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"regexp"
 
-	"github.com/lifthus/froxy/internal/froxyfile"
+	"github.com/lifthus/froxy/internal/config/froxyfile"
 	"github.com/lifthus/froxy/pkg/helper"
 )
