+++ internal/config/froxyfile/testdata/froxyfile
